DevVlog -> DevLog, add update card component w loading states
